backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-cors==3.0.10
tensorflow==2.8.0
numpy>=1.20.0
Pillow==8.3.1
scikit-learn==1.0.2
opencv-python==4.5.3.56
scipy==1.7.1
requests==2.26.0
uuid==1.30 